Refactored Repo
Refactored repository to keep in line with the microservice architecture. In summary, refactored frontend in matching service into communal frontend folder.
