+++ frontend/src/components/CountdownPage.js
@@ -9,7 +9,7 @@ import Box from '@mui/material/Box';
 
 import { useSearchParams } from "react-router-dom"
 import { createSearchParams, useNavigate } from 'react-router-dom';
-import SocketContext from "../context/CreateContext";
+import SocketContext from "./CreateContext";
 
 // Prevent eager initialization of socket
 let socket;
