+++ frontend/src/components/DifficultyPage.js
