+++ frontend/src/components/styles.css
