+++ frontend/src/index.js
@@ -3,10 +3,14 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { SocketProvider } from "./components/CreateContext"
+import { io } from "socket.io-client";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <App/>
+  <SocketProvider value = {io("http://localhost:8001")}>
+    <App />
+  </SocketProvider>
 );
 
 // If you want to start measuring performance in your app, pass a function
